Rename Python package from xc_rbac_sync to xc_user_group_sync (#93)
Complete package rename to better reflect the tool's purpose as a user
and group reconciliation tool, not specifically RBAC management.

**Package Rename:**
- Renamed directory: src/xc_rbac_sync/ → src/xc_user_group_sync/
- Updated all import statements across 9 source files and 13 test files
- Updated CLI entry point in pyproject.toml
- Updated pytest coverage configuration

**Documentation Updates:**
- Updated README.md title and descriptions
- Updated CLAUDE.md project structure
- Updated all specification files (specs/*)
- Updated CI/CD samples (GitHub Actions, Jenkins)
- Updated Serena memory files

**Terminology Cleanup:**
- Replaced "RBAC" references with "user and group synchronization"
- Preserved 3 technically accurate RBAC references (F5 XC API fields)
- Removed 5 misleading RBAC references

**Customer Data Sanitization:**
- Removed customer-specific tenant names (f5-amer-ent, nferreira)
- Replaced with generic examples (acme-corp, acme-staging)
- Updated .claude/settings.local.json, scripts, and specs

**Testing:**
- All 199 tests passing
- All pre-commit hooks passing
- 80%+ code coverage maintained

Fixes #93
