+++ .serena/memories/codebase_structure.md
@@ -4,7 +4,7 @@
 
 ```text
 f5-xc-user-group-sync/
-├── src/xc_rbac_sync/       # Main package
+├── src/xc_user_group_sync/       # Main package
 │   ├── __init__.py         # Package initialization
 │   ├── cli.py              # CLI entry point (Click commands)
 │   ├── client.py           # XCClient - F5 XC API client with retry logic
