+++ .serena/memories/project_overview.md
@@ -1,17 +1,19 @@
-# F5 XC RBAC Group Sync - Project Overview
+# F5 XC User and Group Sync - Project Overview
 
 ## Purpose
 
-Automated synchronization tool for managing F5 Distributed Cloud (XC) RBAC groups from CSV user databases. The tool ensures group membership in F5 XC matches an authoritative user database (e.g., Active Directory exports) with validation, dry-run testing, and automated cleanup capabilities.
+Automated synchronization tool for managing F5 Distributed Cloud (XC) users and groups from CSV user databases. The tool ensures user and group membership in F5 XC matches an authoritative user database (e.g., Active Directory exports) with validation, dry-run testing, and automated cleanup capabilities.
+
+**Key Distinction**: This is a user and group reconciliation tool, not specifically an RBAC management tool. While groups may be used for role-based access control, the tool's primary purpose is to synchronize user and group data from CSV exports to F5 XC.
 
 ## Key Features
 
-- Reads user group memberships from CSV exports (e.g., from Active Directory)
-- Syncs RBAC groups to F5 Distributed Cloud via API
+- Reads user and group memberships from CSV exports (e.g., from Active Directory)
+- Syncs users and groups to F5 Distributed Cloud via API
 - Validates all users exist in XC before creating groups
-- Manages group and user lifecycle (create, update, prune with `--prune`)
+- Manages user and group lifecycle (create, update, prune with `--prune`)
 - Provides dry-run mode for safe testing
-- Integrates with CI/CD pipelines (GitHub Actions)
+- Integrates with CI/CD pipelines (GitHub Actions, Jenkins)
 
 ## Tech Stack
 
@@ -26,12 +28,13 @@ Automated synchronization tool for managing F5 Distributed Cloud (XC) RBAC group
 
 ## Architecture
 
-- **CLI Entry Point**: `xc_rbac_sync.cli` - Command-line interface
-- **API Client**: `xc_rbac_sync.client.XCClient` - F5 XC API interactions with retry logic
-- **Sync Service**: `xc_rbac_sync.sync_service.GroupSyncService` - Business logic for synchronization
-- **Data Models**: `xc_rbac_sync.models` - Pydantic models for validation
-- **Protocols**: `xc_rbac_sync.protocols.GroupRepository` - Dependency injection interface
-- **Utilities**: `xc_rbac_sync.ldap_utils` - LDAP DN parsing
+- **CLI Entry Point**: `xc_user_group_sync.cli` - Command-line interface
+- **API Client**: `xc_user_group_sync.client.XCClient` - F5 XC API interactions with retry logic
+- **Sync Service**: `xc_user_group_sync.sync_service.GroupSyncService` - Business logic for group synchronization
+- **User Sync Service**: `xc_user_group_sync.user_sync_service.UserSyncService` - Business logic for user synchronization
+- **Data Models**: `xc_user_group_sync.models` - Pydantic models for validation
+- **Protocols**: `xc_user_group_sync.protocols.GroupRepository` - Dependency injection interface
+- **Utilities**: `xc_user_group_sync.ldap_utils` - LDAP DN parsing
 
 ## Authentication
