+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "f5-xc-user-group-sync"
 version = "0.1.0"
-description = "F5 XC User and Group Synchronization Tool"
+description = "F5 Distributed Cloud User and Group Synchronization Tool"
 readme = "README.md"
 requires-python = ">=3.9"
 authors = [{ name = "Robin Mordasiewicz" }]
@@ -32,7 +32,7 @@ dev = [
 ]
 
 [project.scripts]
-xc-group-sync = "xc_rbac_sync.cli:cli"
+xc-group-sync = "xc_user_group_sync.cli:cli"
 
 [tool.setuptools]
 include-package-data = false
@@ -66,7 +66,7 @@ addopts = [
   "-ra",
   "--strict-markers",
   "--strict-config",
-  "--cov=xc_rbac_sync",
+  "--cov=xc_user_group_sync",
   "--cov-report=term-missing",
   "--cov-report=html",
   "--cov-fail-under=80",
